cool-common/cool-common-core/src/test/java/it/cnr/cool/cmis/service/VersionServiceTest.java
/*
 * Copyright (C) 2019  Consiglio Nazionale delle Ricerche
 *
 *     This program is free software: you can redistribute it and/or modify
 *     it under the terms of the GNU Affero General Public License as
 *     published by the Free Software Foundation, either version 3 of the
 *     License, or (at your option) any later version.
 *
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU Affero General Public License for more details.
 *
 *     You should have received a copy of the GNU Affero General Public License
 *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package it.cnr.cool.cmis.service;

import it.cnr.cool.MainTestContext;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes = {MainTestContext.class})
public class VersionServiceTest {

    private static final Logger LOGGER = LoggerFactory
            .getLogger(VersionServiceTest.class);
    @Autowired
    private VersionService versionService;

    @Test
    public void testGetVersion() {
        String v = versionService.getVersion();
        LOGGER.info(v);
        assertTrue(v.startsWith("UNKNOWN"));
    }

    @Test
    public void testIsProduction() {
        boolean p = versionService.isProduction();
        LOGGER.info(p ? "produzione" : "sviluppo");
        assertFalse(p);
    }

}
